rearranging functions and bug fixes
