+++ gurobi.log
@@ -44,3 +44,9 @@ Gurobi 9.5.0 (linux64, gurobi_cl) logging started Wed Feb  9 09:52:02 2022
 
 Set parameter Username
 Set parameter LogFile to value "gurobi.log"
+
+Gurobi 9.1.2 (linux64, gurobi_cl) logging started Mon Feb 14 16:01:52 2022
+
+Academic license - for non-commercial use only - expires 2022-03-24
+Using license file /home/user/gurobi.lic
+Set parameter LogFile to value gurobi.log
